backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
chromadb==0.4.18
tiktoken==0.5.2

# PowerPoint and OCR processing
python-pptx==0.6.23
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
aiofiles==23.2.1

# LangChain and LangGraph
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.12
langgraph==0.2.45
python-dotenv==1.1.1

# Document parsing libraries
pdfplumber==0.10.3
python-docx==1.1.0
striprtf==0.0.26
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.4
chardet==5.2.0

# Data.gov API Integration Module Dependencies

# Core HTTP client for async requests
httpx>=0.25.0

# Data manipulation and analysis
pandas>=2.0.0

# Data validation and serialization
pydantic>=2.0.0

# Excel file support for pandas
openpyxl>=3.1.0

# Event loop support for Jupyter notebooks (optional)
nest-asyncio>=1.5.0

# Type checking (development)
mypy>=1.5.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
